Separated the API calls in different files.
